+++ test/revisionGraphHostMessages.test.ts
@@ -60,7 +60,8 @@ function createReadyRevisionGraphState(): RevisionGraphViewState {
       showRemoteBranches: true,
       showStashes: true,
       showMergeCommits: false,
-      showCurrentBranchDescendants: false
+      showCurrentBranchDescendants: false,
+      revisionRange: undefined
     },
     mergeBlockedTargets: [],
     primaryAncestorNextByHash: {},